ct: Move 'ctp_stm_key' to the stm/types.h
