+++ src/v/cloud_topics/level_zero/stm/types.h
@@ -16,6 +16,10 @@
 
 namespace experimental::cloud_topics {
 
+enum class ctp_stm_key : uint8_t {
+    advance_reconciled_offset = 1,
+};
+
 struct [[nodiscard]] cluster_epoch_fence {
     // Units protecting the epoch state.
     // If it's set to nullopt the batch has to be discarded
@@ -25,3 +29,11 @@ struct [[nodiscard]] cluster_epoch_fence {
     model::term_id term;
 };
 } // namespace experimental::cloud_topics
+
+template<>
+struct fmt::formatter<experimental::cloud_topics::ctp_stm_key>
+  : fmt::formatter<std::string_view> {
+    auto format(
+      experimental::cloud_topics::ctp_stm_key, fmt::format_context& ctx) const
+      -> decltype(ctx.out());
+};
