--- src/v/cloud_topics/tests/types_test.cc
@@ -1,18 +1,0 @@
-/*
- * Copyright 2024 Redpanda Data, Inc.
- *
- * Licensed as a Redpanda Enterprise file under the Redpanda Community
- * License (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
- */
-#include "cloud_topics/types.h"
-
-#include <gtest/gtest.h>
-
-TEST(DlStmKey, Formatting) {
-    ASSERT_EQ(
-      fmt::format("{}", experimental::cloud_topics::ctp_stm_key(255)),
-      "unknown ctp_stm_key(255)");
-}
